requirements.txt
python-dotenv~=1.0.1
requests~=2.29.0
httpx~=0.27.0
openai~=1.16.2